requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Authentication & Security
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Image Processing
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
bleach==6.1.0

# Production Server (Optional)
gunicorn==21.2.0

# Testing & Development
pytest==7.4.3
pytest-flask==1.3.0
